Turned class Packr into module and extracted parts of old Packr class into Packr::Engine. Reason? The previous implementation makes it "impossible" to extend the Packr namespace, e.g. "Packr::Rails" (a.k.a. "Packr is not a class"). Also this is the better convention (see Bundler, etc.), and wrapping class in class makes no sense anyway.
