+++ lib/packr.rb
@@ -9,24 +9,25 @@
   '/packr/privates',
   '/packr/shrinker',
   '/packr/words',
-  '/packr/base62'
+  '/packr/base62',
+  '/packr/engine'
 ].each do |path|
   require File.dirname(__FILE__) + path
 end
 
-class Packr
-  
+module Packr
+
   DATA = Parser.new.
     put("STRING1", IGNORE).
     put('STRING2', IGNORE).
     put("CONDITIONAL", IGNORE). # conditional comments
     put("(OPERATOR)\\s*(REGEXP)", "\\1\\2")
-  
+
   def self.encode62(c)
     (c < 62 ? '' : encode62((c / 62.0).to_i)) +
         ((c = c % 62) > 35 ? (c+29).chr : c.to_s(36))
   end
-  
+
   def self.encode52(c)
     # Base52 encoding (a-Z)
     encode = lambda do |d|
@@ -37,26 +38,11 @@ class Packr
     encoded = encoded[1..-1] + '0' if encoded =~ /^(do|if|in)$/
     encoded
   end
-  
+
   def self.pack(script, options = {})
-    @packr ||= self.new
+    @packr ||= Packr::Engine.new
     @packr.pack(script, options)
   end
-  
-  def initialize
-    @minifier = Minifier.new
-    @shrinker = Shrinker.new
-    @privates = Privates.new
-    @base62   = Base62.new
-  end
-  
-  def pack(script, options = {})
-    script = @minifier.minify(script)
-    script = @shrinker.shrink(script, options[:protect]) if options[:shrink_vars]
-    script = @privates.encode(script) if options[:private]
-    script = @base62.encode(script) if options[:base62]
-    script
-  end
-  
+
 end
