+++ lib/packr/map.rb
@@ -1,4 +1,4 @@
-class Packr
+module Packr
   # This is effectively a wrapper for Hash instances - we're including it
   # to maintain similarity with the JavaScript version for easier maintainance.
   class Map
