+++ lib/packr/regexp_group.rb
@@ -1,4 +1,4 @@
-class Packr
+module Packr
   class RegexpGroup < Collection
     
     IGNORE          = "\\0"
