+++ test/test_packr.rb
@@ -6,7 +6,7 @@ require 'fileutils'
 require File.expand_path('../../lib/packr', __FILE__)
 
 class PackrTest < Test::Unit::TestCase
-  
+
   def setup
     dir = File.dirname(__FILE__) + '/assets'
     FileUtils.mkdir_p(dir + '/test')
@@ -47,13 +47,13 @@ class PackrTest < Test::Unit::TestCase
       }]
     }
   end
-  
+
   def test_basic_packing
     actual = Packr.pack(@data[:default][0][:source])
     File.open(@data[:default][0][:output], 'wb') { |f| f.write(actual) }
     assert_equal @data[:default][0][:packed], actual
   end
-  
+
   def test_shrink_packing
     actual1 = Packr.pack(@data[:shrink][0][:source], :shrink_vars => true)
     File.open(@data[:shrink][0][:output], 'wb') { |f| f.write(actual1) }
@@ -62,7 +62,7 @@ class PackrTest < Test::Unit::TestCase
     assert_equal @data[:shrink][0][:packed], actual1
     assert_equal @data[:shrink][1][:packed], actual2
   end
-  
+
   def test_base62_packing
     expected = @data[:base62][0][:packed]
     actual = Packr.pack(@data[:base62][0][:source], :base62 => true)
@@ -71,8 +71,8 @@ class PackrTest < Test::Unit::TestCase
     expected_words = expected.scan(/'[\w\|]+'/)[-2].gsub(/^'(.*?)'$/, '\1').split("|").sort
     actual_words = actual.scan(/'[\w\|]+'/)[-2].gsub(/^'(.*?)'$/, '\1').split("|").sort
     assert expected_words.eql?(actual_words)
-  end  
-  
+  end
+
   def test_base62_and_shrink_packing
     expected = @data[:base62_shrink][0][:packed]
     actual = Packr.pack(@data[:base62_shrink][0][:source], :base62 => true, :shrink_vars => true)
@@ -82,13 +82,13 @@ class PackrTest < Test::Unit::TestCase
     actual_words = actual.scan(/'[\w\|]+'/)[-2].gsub(/^'(.*?)'$/, '\1').split("|").sort
     assert expected_words.eql?(actual_words)
   end
-  
+
   def test_private_variable_packing
     script = "var _KEYS = true, _MY_VARS = []; (function() { var foo = _KEYS;  _MY_VARS.push({_KEYS: _KEYS}); var bar = 'something _KEYS  _MY_VARS' })();"
     expected = "var _0=true,_1=[];(function(){var a=_0;_1.push({_0:_0});var b='something _0  _1'})();"
     assert_equal expected, Packr.pack(script, :shrink_vars => true, :private => true)
   end
-  
+
   def test_protected_names
     expected = 'var func=function(a,d,c,b){return c(a+b)}'
     actual = Packr.pack('var func = function(foo, bar, $super, baz) { return $super( foo + baz ); }', :shrink_vars => true)
@@ -99,41 +99,41 @@ class PackrTest < Test::Unit::TestCase
     expected = 'function(a,$super){}'
     assert_equal expected.size, Packr.pack('function(name, $super) { /* something */ }', :shrink_vars => true, :protect => %w($super)).size
   end
-  
+
   def test_dollar_prefix
     expected = 'function(a,b){var c;happening()}'
     actual = Packr.pack('function(something, $nothing) { var is; happening(); }', :shrink_vars => true)
     assert_equal expected, actual
   end
-  
+
   def test_object_properties
     expected = 'function(a,b){this.queue.push({func:a,args:b})}'
     actual = Packr.pack('function(method, args) { this.queue.push({func: method, args: args}); }', :shrink_vars => true)
     assert_equal expected, actual
   end
-  
+
   def test_concat
     actual = Packr.pack(@data[:concat_bug][0][:source], :shrink_vars => true)
     File.open(@data[:concat_bug][0][:output], 'wb') { |f| f.write(actual) }
     assert_equal @data[:concat_bug][0][:packed], actual
-    
+
     code = 'var a={"+":function(){}}'
     assert_equal Packr.pack(code), code
-    
+
     code = "var a={'+':function(){}}"
     assert_equal Packr.pack(code), code
-    
+
     code = 'var b="something"+"else",a={"+":function(){return"nothing"+"at all"}}'
     expected = 'var b="somethingelse",a={"+":function(){return"nothingat all"}}'
     actual = Packr.pack(code)
     assert_equal expected, actual
-    
+
     code = "var b='something'+'else',a={'+':function(){return'nothing'+'at all'}}"
     expected = "var b='somethingelse',a={'+':function(){return'nothingat all'}}"
     actual = Packr.pack(code)
     assert_equal expected, actual
   end
-  
+
   def test_conditional_comments
     expected = @data[:conditional_comments][0][:packed]
     actual = Packr.pack(@data[:conditional_comments][0][:source], :shrink_vars => true)
